requirements.txt
asyncio-mqtt>=0.16.1
suta-ble-bed>=0.2.0
PyYAML>=5.3
